+++ src/internal/puck/components/ThemeSidebar.tsx
@@ -1,7 +1,66 @@
+import { AutoFieldPrivate } from "@measured/puck";
 import React from "react";
+import { Alert, AlertDescription } from "../../components/atoms/Alert.tsx";
+import { ThemeConfig } from "../../../utils/themeResolver.ts";
+import {
+  constructThemePuckFields,
+  constructThemePuckValues,
+} from "../../utils/constructThemePuckFields.ts";
 
-const ThemeSidebar = () => {
-  return <div></div>;
+type ThemeSidebarProps = {
+  savedThemeValues: any;
+  themeConfig?: ThemeConfig;
+  saveTheme: (theme: ThemeConfig) => void;
+};
+
+const ThemeSidebar = (props: ThemeSidebarProps) => {
+  const { saveTheme, themeConfig, savedThemeValues } = props;
+  if (!themeConfig) {
+    return (
+      <div>
+        <Alert>
+          <AlertDescription>
+            Please contact your developer to set up theme management.
+          </AlertDescription>
+        </Alert>
+      </div>
+    );
+  }
+
+  const handleChange = (themeCategoryKey: string, newValue: any) => {
+    const updatedConfig: ThemeConfig = {
+      ...themeConfig,
+      [themeCategoryKey]: {
+        ...themeConfig[themeCategoryKey],
+        styles: newValue,
+      },
+    };
+    saveTheme(updatedConfig);
+  };
+
+  return (
+    <div>
+      <Alert>
+        <AlertDescription>
+          Changes to the theme will affect all templates and layouts for the
+          entire site.
+        </AlertDescription>
+      </Alert>
+      {Object.entries(themeConfig).map(([parentStyleKey, parentStyle]) => {
+        const field = constructThemePuckFields(parentStyle);
+        const values = constructThemePuckValues(savedThemeValues, parentStyle);
+
+        return (
+          <AutoFieldPrivate
+            key={parentStyleKey}
+            field={field}
+            onChange={(value) => handleChange(parentStyleKey, value)}
+            value={values}
+          />
+        );
+      })}
+    </div>
+  );
 };
 
 export default ThemeSidebar;